factored out common logic between BusinessList and AgentList to list
