+++ scripts/agents/AgentList.js
@@ -1,3 +1,4 @@
+import { renderListToDom, createListArticleNode } from '../list/list.js';
 import { useAgents, useAgentsMatchingPropertyValue } from './AgentProvider.js';
 import { Agent } from './Agent.js';
 
@@ -7,25 +8,8 @@ const container = document.querySelector('.container');
 eventHub.addEventListener('searched', event => {
   const { searchTerm } = event.detail;
 
-  // renderMatchingAgents(searchTerm);
-
   AgentListWhere('agents--found', 'Search Results', 'fullName', searchTerm);
 });
-
-/**
- * 
- * @param {HTMLElement} domNode A reference to the DOM node that you want to add this list's HTML to.
- * @param {Array} agents An array of agent objects.
- * @param {String} listHeading The heading / title that you want for this list.
- */
-const renderListToDom = (domNode, agents, listHeading) => {
-  const agentsHTML = agents.map(Agent).join('\n');
-
-  domNode.innerHTML = `
-    <h2 class="list__heading">${listHeading}</h2>
-    ${agentsHTML}
-  `; 
-};
 
 export const AgentListWhere = (className, heading, property, value) => {
   let domNode = document.querySelector(`.${className}`);
@@ -36,18 +20,7 @@ export const AgentListWhere = (className, heading, property, value) => {
 
   const agents = getAgentsList(property, value);
 
-  renderListToDom(domNode, agents, heading);
-};
-
-/**
- * Create and return an <article> DOM node with classes 'list' and className
- * @param {String} className The custom class name to give the element
- */
-const createListArticleNode = className => {
-  const domNode = document.createElement('ARTICLE');
-  domNode.classList.add('list');
-  domNode.classList.add(className);
-  return domNode;
+  renderListToDom(domNode, agents, Agent, heading);
 };
 
 /**
